move chat helpers out of protocol
